move avatar outside of users dir
