+++ app/components/avatar_component.html.erb
