+++ app/components/avatar_component.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+class AvatarComponent < ViewComponent::Base
+  def initialize(user:, size:)
+    @user = user
+    @size = size
+  end
+
+  private
+
+  attr_reader :user, :size
+end
