+++ app/views/comments/_comment.html.erb
@@ -1,5 +1,5 @@
 <div class="flex justify-start items-start gap-4 w-full ">
-  <%= render Users::AvatarComponent.new(user: comment.user, size: 8) %>
+  <%= render AvatarComponent.new(user: comment.user, size: 8) %>
   <div class="flex flex-col justify-start items-start truncate">
     <div class="flex justify-start items-center gap-1 text-secondary-txt text-[0.6rem] font-light">
       <p><%= comment.user.display_name %></p>
