+++ app/views/nav/_navbar.html.erb
@@ -56,7 +56,7 @@
       <li>
         <button id="navbar-user-dropdown-btn" type="button" class="nav-dropdown"
           data-dropdown-toggle="navbar-user-dropdown" data-dropdown-placement="bottom-start">
-          <%= render Users::AvatarComponent.new(user: current_user, size: 6) %>
+          <%= render AvatarComponent.new(user: current_user, size: 6) %>
           <%= icon "chevron-down", class: "w-3 h-3 min-w-3 min-h-3 aspect-square text-secondary-txt" %>
         </button>
