+++ spec/components/admin/track_spec.rb
